+++ resources/views/container/systems/category/grid.blade.php
@@ -11,7 +11,7 @@
     <div class="text-right">
         <div class="btn-group" role="group">
             <a href="{{ route('dashboard.systems.category.create')}}" class="btn btn-link"><i
-                        class="icon-plus fa fa-2x"></i></a>
+                        class="sli icon-plus fa-2x"></i></a>
         </div>
     </div>
 @stop