+++ resources/views/fields/input.blade.php
@@ -2,7 +2,7 @@
     @if(isset($title))
         <label for="field-{{$slug}}">{{$title}}</label>
     @endif
-    <input type="{{$type}}" class="form-control {{$class or ''}}" id="field-{{$slug}}"
+    <input type="{{$type}}" class="form-control {{$class or ''}}" id="field-{{$lang}}-{{$slug}}"
            @if(isset($prefix))
            name="{{$prefix}}[{{$lang}}]{{$name}}"
            @else